# Remove exponential-time Levenshtein from assembler